fix(test): split diplomatic order tests to satisfy logic test file size gate
Refs #2504
